slicer: Fix output logging
std::stringstream(std::string& x) initializes the buffer with x, but
does not update the output position. The recent change that added this
yielded error messages with overwritten content.

Flush the log buffer since logging is to stdout by default and
buffered so log messages can be swallowed if abort is called.

Bug: N/A
Test: dexter_tests.py & slicer_tests & manual
Change-Id: I0587115f6eff495c54f5073bbb8723b587db3242